+++ changelog.md
@@ -9,6 +9,7 @@ Changelog
 * Admins can delete users
 * Decaching for CSS and JavaScript assets
 * Folder validator now check for special characters
+* Users can be promoted to admins
 
 1.13.2
 ------